Save system final fix (I hope)

+++ Assets/Scripts/SaveSystem/SaveSystem.cs
@@ -6,13 +6,18 @@ public class SaveSystem
     public static void SaveData(SaveableDatas datas)
     {
         string path = Application.persistentDataPath + $"/{datas.GetFileName()}.wad";
-
         string dataToStore = JsonUtility.ToJson(datas, true);
+
+        if (File.Exists(path))
+        {
+            File.WriteAllText(path, dataToStore );
+            return;
+        }
 
         FileStream stream = new FileStream(path, FileMode.Create);
         StreamWriter writer = new StreamWriter(stream);
         writer.Write(dataToStore);
-
+        writer.Close();
         stream.Close();
     }
